all sorts of madness (re: factual integration)
basically we now require node modules, including the factual node module

this runs on the server since factual can't auth from the client - this also helps us guarantee only our logged in customers go on our factual bill (we might need similar things for mapbox eventually).

this also does async calls with meteor methods, which is not for the faint of heart, but it worked as documented and wasn't too bad.  anyway upshot is we can now run factual queries from the client.  can't visualize em just yet, but hey.