up- and down-voting was surprisingly easy
can now up and down vote collections and places.

note that I don't actually do anything with the votes that it will be trivial to sort by number of votes if we every get enough votes that we trust vote counts.  for now i'm going to leave it as sorting by most recent first for this reason.

I've added voting for collections and places.  It would be easy to add it for posts or comments too, but that seems like overkill right now.