allow import of flickr data
(not the actual images)

this lets you edit the name and description for all the images, which you can't currently do by getting the flickr data on-the-fly.  Technically with the old way you need to edit the caption and description on Flickr directly.  Is this way better slash should we remove the old way?